Quick update
If a user's browser contains a userhash cookie, but no attached party is found, the cookie is cleared and user redirected to index.

Fixed the existing party check. if the spotify user ID already has an open party but has lost their token, they will be redirected to their old party and provided a new token.